LIBFCREPO-1126: Moving is_handle function to validation
https://umd-dit.atlassian.net/browse/LIBFCREPO-1146
