+++ plastron/models/umd.py
@@ -2,8 +2,7 @@ from plastron import pcdm, rdf
 from plastron.authority import LabeledThing
 from plastron.namespaces import dc, dcterms, edm
 from plastron.pcdm import Page
-from plastron.validation import is_edtf_formatted, is_valid_iso639_code
-from plastron.util import is_handle
+from plastron.validation import is_edtf_formatted, is_valid_iso639_code, is_handle
 from rdflib import Namespace
